-add AutorizationHandler, remove table employe

+++ LibraryApi/Services/LibraryService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
+using LibraryApi.Authorization;
 using LibraryApi.Entieties;
 using LibraryApi.Exceptions;
 using LibraryApi.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace LibraryApi.Services
@@ -15,27 +18,36 @@ namespace LibraryApi.Services
     {
         private readonly LibraryDbContext _dbContext;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
 
-        public LibraryService(LibraryDbContext dbContext, IMapper mapper)
+        public LibraryService(LibraryDbContext dbContext, IMapper mapper, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
-        public int Create(CreateLibraryDto dto)
+        public int Create(CreateLibraryDto dto, int userId)
         {
             var library = _mapper.Map<Library>(dto);
+            library.CreatedById = userId;
             _dbContext.Add(library);
             _dbContext.SaveChanges();
             return library.Id;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             var library = _dbContext.Libraries.FirstOrDefault(l => l.Id == id);
             if(library == null)
             {
                 throw new NotFoundException("Library not found");
+            }
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, library,
+                new ResourceOperationRequirement(ResourceOperation.DELETE)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
             }
             _dbContext.Libraries.Remove(library);
             _dbContext.SaveChanges();
